+++ akshare/__init__.py
@@ -2019,9 +2019,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.5.97 fix: fix stock_zh_a_hist interface
 1.5.98 fix: fix news_cctv interface
 1.5.99 fix: fix stock_zh_a_hist_163 interface
+1.6.1 fix: fix stock_info_sh_name_code interface
 """
 
-__version__ = "1.5.99"
+__version__ = "1.6.1"
 __author__ = "AKFamily"
 
 import sys