+++ docs/changelog.md
@@ -29,6 +29,10 @@
 
 ## 更新说明
 
+1.6.1 fix: fix stock_info_sh_name_code interface
+
+    1. 修复 stock_info_sh_name_code 接口，字段对齐
+
 1.5.99 fix: fix stock_zh_a_hist_163 interface
 
     1. 修复 stock_zh_a_hist_163 接口，修改编码格式为 GBK 编码
@@ -681,6 +685,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.6.1 fix: fix stock_info_sh_name_code interface
 
 1.5.99 fix: fix stock_zh_a_hist_163 interface